+++ src/Bridge/Symfony/DependencyInjection/SeoOverrideExtension.php
@@ -28,6 +28,10 @@ class SeoOverrideExtension extends Extension
 
         $this->registerFetchersConfiguration($config['fetchers'], $container, $loader);
         $this->registerDomainsConfiguration($config['domains'], $container);
+
+        if ($container->hasParameter('kernel.debug') && $container->getParameter('kernel.debug')) {
+            $loader->load('debug.yml');
+        }
     }
 
     private function registerFetchersConfiguration(array $fetchers, ContainerBuilder $container, YamlFileLoader $loader)