+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes between versions
 
+## Unreleased
+
+* Add Web Profiler data collector
+
 ## 0.1.2 (2017-05-12)
 
 * Fix return typehint on Doctrine entity getters